# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeForge AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories:
    get:
      operationId: listRepositories
      tags: [repositories]
      summary: List connected repositories
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
    post:
      operationId: connectRepository
      tags: [repositories]
      summary: Connect a new repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectRepositoryBody"
      responses:
        "201":
          description: Repository connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryDetail"
  /repositories/{id}/scan:
    post:
      operationId: scanRepository
      tags: [repositories]
      summary: Trigger a repository scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
  /repositories/{id}/graph:
    get:
      operationId: getRepositoryGraph
      tags: [repositories]
      summary: Get the code knowledge graph for a repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository graph nodes and edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryGraph"
...
  /executions:
    get:
      operationId: listExecutions
      tags: [executions]
      summary: List terminal executions
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
    post:
      operationId: createExecution
      tags: [executions]
      summary: Run a terminal command (sandboxed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExecutionBody"
      responses:
        "201":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /executions/{id}:
    get:
      operationId: getExecution
      tags: [executions]
      summary: Get execution details and output
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Execution details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
...
components:
  schemas:
...
    Repository:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fullName:
          type: string
        provider:
          type: string
          enum: [github, gitlab, bitbucket, azure, local]
        url:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [connected, scanning, ready, error]
        lastScannedAt:
          type: string
          nullable: true
        fileCount:
          type: integer
        lineCount:
          type: integer
        frameworks:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, name, fullName, provider, url, language, status, fileCount, lineCount, frameworks, createdAt]
    RepositoryDetail:
      allOf:
        - $ref: "#/components/schemas/Repository"
        - type: object
          properties:
            description:
              type: string
            branches:
              type: array
              items:
                type: string
            dependencies:
              type: array
              items:
                type: string
            apis:
              type: array
              items:
                type: string
            databases:
              type: array
              items:
                type: string
            cicdPlatform:
              type: string
              nullable: true
          required: [description, branches, dependencies, apis, databases]
    ConnectRepositoryBody:
      type: object
      properties:
        url:
          type: string
        provider:
          type: string
          enum: [github, gitlab, bitbucket, azure, local]
        name:
          type: string
      required: [url, provider, name]
    ScanResult:
      type: object
      properties:
        repositoryId:
          type: string
        status:
          type: string
          enum: [started, running, complete, error]
        filesScanned:
          type: integer
        issuesFound:
          type: integer
        startedAt:
          type: string
      required: [repositoryId, status, filesScanned, issuesFound, startedAt]
    RepositoryGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
      required: [nodes, edges]
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [function, class, api, service, database, file]
        file:
          type: string
      required: [id, label, type, file]
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        relation:
          type: string
          enum: [calls, extends, imports, writes, reads]
      required: [source, target, relation]
...
    Execution:
      type: object
      properties:
        id:
          type: string
        command:
          type: string
        status:
          type: string
          enum: [queued, running, success, error, blocked]
        output:
          type: string
        exitCode:
          type: integer
          nullable: true
        sessionId:
          type: string
          nullable: true
        durationMs:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, command, status, output, createdAt]
    CreateExecutionBody:
      type: object
      properties:
        command:
          type: string
        sessionId:
          type: string
          nullable: true
      required: [command]
...