# lib/api-zod/src/generated/types/graphEdge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeForge AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GraphEdgeRelation } from "./graphEdgeRelation";

export interface GraphEdge {
  source: string;
  target: string;
  relation: GraphEdgeRelation;
}