# lib/api-zod/src/generated/types/graphNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeForge AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GraphNodeType } from "./graphNodeType";

export interface GraphNode {
  id: string;
  label: string;
  type: GraphNodeType;
  file: string;
}